Lots of code *cleanup (part 2)

+++ modules/core/include/argus/core/callback.hpp
@@ -9,6 +9,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 namespace argus {
-    typedef unsigned long long Index;
+    typedef uint64_t Index;
 }
